+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - New feature to find all possible solutions instead of just one.
 - New feature of multiple shop system to meet mandatory original Constraint Number 3
+- New features for robust menu system with consistent looping feature until the user decides to exit.
+- New feature to allow user to add custom constraints
+- New feature to allow user to add constraints based on person, shop type, and items s type
+- New feature to allow user to view initial default and newly introduced constraints
+- New feature to set solutions display options by person, shop type, and items type
+- New feature to allow user to find all or specific solutions by person, shop type, or / and items type.
+- New feature to allow user to exit and catch keyboard interruption gracefully.
 
 ### Changed
 - Improved the code readability with more consistent for Bobby's constraints.